Restore most gems

+++ db/migrate/20200311011304_remove_product_orders.rb
@@ -1,5 +1,5 @@
 class RemoveProductOrders < ActiveRecord::Migration
   def change
-    drop_table :product_orders
+    drop_table :product_orders if table_exists?(:product_orders)
   end
 end
