+++ db/migrate/20200311011643_remove_products.rb
@@ -1,5 +1,5 @@
 class RemoveProducts < ActiveRecord::Migration
   def change
-    drop_table :products
+    drop_table :products if table_exists?(:products)
   end
 end
